+++ src/stat.c
@@ -7,26 +7,53 @@
 #include <time.h> /* nanosleep, timespec */
 #include <errno.h> /* errono, strerror */
 
+#include "info.h"
 #include "error.h"
 
 #define PROGRAM_NAME "stat"
+#define PROGRAM_DESC "Output file status"
+
+const char *usages[] = {
+	"FILE..."
+};
+
+t_arg_desc arg_desc[] = {
+	{"--help",    "Show command help"},
+	{"--version", "Show tcore version"}
+};
 
 void get_stat_date(char *p_buf, size_t p_size, time_t *p_sec) {
 	strftime(p_buf, p_size, "%y-%m-%d", localtime(p_sec));
 }
 
 int main(int p_argc, const char **p_argv) {
-	if (p_argc == 2 && strcmp(p_argv[1], "-v") == 0) {
-		printf(
-			"tcore %s version %i.%i.%i\n",
-			PROGRAM_NAME,
-			VERSION_MAJOR,
-			VERSION_MINOR,
-			VERSION_PATCH
-		);
-		return EXIT_SUCCESS;
+	if (p_argc == 2) {
+		if (strcmp(p_argv[1], "--version") == 0) {
+			version(PROGRAM_NAME, VERSION_MAJOR, VERSION_MINOR, VERSION_PATCH);
+			return EXIT_SUCCESS;
+		} else if (strcmp(p_argv[1], "--help") == 0) {
+			if (
+				help(
+					PROGRAM_NAME,
+					usages, sizeof(usages) / sizeof(const char*),
+					PROGRAM_DESC,
+					arg_desc, sizeof(arg_desc) / sizeof(t_arg_desc)
+				) != EXIT_SUCCESS
+			)
+				error_fatal(PROGRAM_NAME);
+			else
+				return EXIT_SUCCESS;
+		}
 	}
 	ERR_WATCH_INIT;
+
+	if (p_argc < 2) {
+		ERR_SET_G_ERROR("Missing argument", NULL, ERR_NOT_FATAL);
+		error_simple(PROGRAM_NAME);
+		error_cleanup();
+		try(PROGRAM_NAME);
+		return EXIT_FAILURE;
+	}
 
 	int exitcode = EXIT_SUCCESS;
 	for (++ p_argv; *p_argv != NULL; ++ p_argv) {